cifar/requirements.txt
tensorflow
numpy
flask
gunicorn
pillow
